+++ src/sc/blog/browser/blog.py
@@ -6,6 +6,7 @@ from plone import api
 from plone.app.layout.viewlets.interfaces import IAboveContent
 from plone.directives import dexterity
 from Products.CMFCore.interfaces import ISiteRoot
+from sc.blog.config import BLOG_BLACKLISTED_TYPES
 from sc.blog.content import IBlog
 from sc.blog.interfaces import IBlogLayer
 from zope.interface import Interface
@@ -21,16 +22,17 @@ class View(dexterity.DisplayForm):
     grok.template('folder_full_view')
     grok.require('zope2.View')
 
-    def query_portal_types(self):
-        """ Returns the portal types than should be displayed
-            as post entries inside a Blog.
-            'Image', 'File', 'Folder' & 'Blog' are excluded.
+    def get_blog_friendly_types(self):
+        """Return the portal types than should be displayed as post entries
+        inside a Blog. Some content types (like Blog, File, Folder and Image)
+        are excluded by default.
+
+        :returns: a tuple with the names of the portal types
         """
         plone_utils = api.portal.get_tool(name='plone_utils')
-        types = plone_utils.getUserFriendlyTypes()
-        for t in ('Image', 'File', 'Folder', 'Blog'):  # XXX: hardcoded, please improve
-            types.remove(t)
-        return types
+        friendly_types = plone_utils.getUserFriendlyTypes()
+        blog_types = set(friendly_types) - set(BLOG_BLACKLISTED_TYPES)
+        return tuple(blog_types)
 
 
 class BlogSummaryView(View):