+++ src/sc/blog/config.py
@@ -1,3 +1,11 @@
 # -*- coding: utf-8 -*-
 
 PROJECTNAME = 'sc.blog'
+
+# content types that can not be considered as blog posts by default
+BLOG_BLACKLISTED_TYPES = (
+    'Blog',
+    'File',
+    'Folder',
+    'Image',
+)